src/main/python/requirements.txt
torch==1.6.0
imgaug==0.4.0
numpy==1.19.1
scikit-image==0.17.2
tqdm==4.48.0
torchvision==0.7.0
segmentation-models-pytorch==0.1.2
pytorch-ignite>=0.4.2